Change AppVeyor to test more versions, no wheels
We already generate wheels in our GitHub Actions via cibuildwheel, so
there's no need to generate them in AppVeyor anymore (we used to when we
manually generated wheels for different platforms). On the flip side, we
can test against more python versions, and only focus on 64-bit builds.

Keeping AppVeyor around is still a good thing, as it gives us quick
feedback if there's something broken with a build on Windows without the
overheads of cibuildwheel.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>