Add an environment variable to control building yajl2_c
Add a IJSON_BUILD_YAJL2C environment variable to make it possible
to override whether yajl2_c is being built.  The default is to build it
if either yajl2 is found or embedding is enabled.  However, the envvar
can be used to either force building (i.e. effectively make the build
fail if yajl2 is not detected) or disable building it even if yajl2
is installed.

The primary use case is to make it possible to disable building
the extension for PyPy3 where it is currently known to be broken.

Signed-off-by: Michał Górny <mgorny@gentoo.org>