Rename and correctly size StopIteration args
When returning a value via StopIteration.value, we need to provide this
single value as a 1-tuple. We were correctly setting the 0-th value in
the tuple, but we were actually creating it incorrectly, as it had size
2 instead of 1. CPython dealt with this bug without an issue, but when
running against PyPy an error was raised since the second element in the
tuple was NULL.

This commit correctly sizes the tuple to be of size 1 instead of 2.
Moreover, it renames the variable from "res" (not very useful) to
"stop_iteration_args", which better captures its meaning.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>