Raise exception after calling DelSlice
When raising a StopIteration exception to return a value for our
coroutine, we were raising it before calling PySequence_DelSlice. This
was incorrect, because one should not invoke further python functions
when an exception has been set, otherwise those further invocations
would return non-NULL values with an exception set. Normal python builds
don't care about this, but python debug builds catch these errors, which
can be otherwise difficult to debug.

This commit re-organises our code so we clear our internal list *before*
we raise the exception, so there are no further Python functions invoked
after the exception is set.

This issue was found while debugging PyPy crashes in #101, even though
it's not necessarily the case that those crashes were caused by this
problem.