Build StopIteration exception fully
When raising exceptions from a C extension via PyErr_SetObject(type,
value), they can be normalised or not. In the normalised case, "value"
is an instance of the exception type (e.g., "value =
StopIteration(args)"), while the latter simply passes the arguments that
would be given when constructing the exception type instance, without
constricting it (following the example above, "value = args").

Historically we have had no issues with raising non-normalised
exceptions. However, this doesn't currently work with PyPy (reported in
https://foss.heptapod.net/pypy/pypy/-/issues/3965). Raising normalised
exceptions *does* work though, so we choose to do that instead.

Note that contrary to a comment in that issue, performance *is* impacted
for CPython if we choose to always raise normalised exceptions (local
tests show ~10-20% degradation, depending on the test case). We
therefore do this only when building against PyPy, leaving CPython
working as previously.

This issue was spotted in #101.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>