Fix refcount and make it more intuitive
The chain function in coro_utils was special casing some reference
counting to avoid touching the "sink" argument's reference count, while
at the same time "sink" was the value that the internal "coro" object
was initialised to. This special casing made it difficult to understand
the correctness of the reference counts of "coro" and "sink", which has
led to small refcount bugs over the years.

This commit changes how to handle "sink" entirely by properly owning it
for the duration of the call. Instead of leaving its refcount untouched,
we increase its refcount by one after assigning it to "coro"; after
that, we only care about "coro"'s refcount correctness.

Moreover, this commit also fixes a refcount issue: we were passing
"coro" to PyTuple_SET_ITEM, which steals a reference, without
incrementing "coro"'s refcount ourselves. This issue is in a branch of
the code that ultimately gets triggered only by the "kvitems" and
"items" functions, and in both cases "coro" was the same as "sink",
which in turn corresponded to the internal Py_List object build by the
kvitems/items generators and async generators. Thus, this issue
manifested itself as a "list"-related crash in pypy, and as a "negative
refcount found" for an object of type "list" when running with CPython
debug builds.

This issue was found as part of the crash debugging done in #101.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>